+++ chaos/common/direct_io/channel/DirectIOCDataWrapperClientChannel.h
@@ -39,7 +39,7 @@ namespace chaos {
                      DirectIOCDataWrapperClientChannel(std::string alias);
                  public:
 					~DirectIOCDataWrapperClientChannel();
-                    uint32_t pushCDataWrapperSerializationBuffer(uint16_t endpoint_idx, uint8_t opcode, chaos_data::SerializationBuffer *data_pack);
+                    int64_t pushCDataWrapperSerializationBuffer(uint16_t endpoint_idx, uint8_t opcode, chaos_data::SerializationBuffer *data_pack);
                  };
             }
         }
